refactor: move `users` to TS
