+++ package.json
@@ -56,6 +56,7 @@
   "devDependencies": {
     "@types/express": "^4.17.11",
     "@types/swagger-ui-express": "^4.1.2",
+    "@types/uuid": "^8.3.0",
     "@types/yamljs": "^0.2.31",
     "@typescript-eslint/eslint-plugin": "^4.25.0",
     "chai": "4.3.4",
