--- src/resources/users/user.memory.repository.js
@@ -1,3 +1,0 @@
-const MemoryRepository = require('../../common/memory.repository');
-
-module.exports = new MemoryRepository();
