+++ src/resources/users/user.memory.repository.ts
@@ -1,0 +1,5 @@
+import MemoryRepository from 'common/memory.repository';
+
+import { IUser } from './user.model';
+
+export default new MemoryRepository<string, IUser>();
